+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add `EveOnline::ESI::MarketHistory`
 * [Sync with ESI 2017/12/20 update](https://developers.eveonline.com/blog/article/esi-endpoint-versioning-important-info-and-best-practices)
 * Add `EveOnline::ESI::CharacterOnline`
+* Add `EveOnline::ESI::Models::Online`
 
 **v0.13.0**